[Backend] 광고/펌웨어 배포 완료 시 광고/펌웨어 히스토리 저장 로직 추가 (#160)
* fix: download_seconds 타입을 float에서 long으로 변경

* feat: 광고 배포 완료 시 device_ads 테이블 업데이트 로직 추가

* feat: 펌웨어 배포 완료 시 device_ads 테이블 업데이트 로직 추가

+++ backend/iot-cloud-ota/src/main/java/com/coffee_is_essential/iot_cloud_ota/entity/FirmwareDownloadEvents.java
@@ -17,6 +17,6 @@ public class FirmwareDownloadEvents {
     private Long downloadBytes;
     private double speedKbps;
     private boolean checksumVerified;
-    private double downloadTime;
+    private Long downloadTime;
     private Timestamp timestamp;
 }
